Changed to work with our IDs and added arm code
I added in all the encoder and motor IDs as well as changed it to be  Compatible with our hardware. I also added a arm file that is the equivalent of the drivetrain + swerveModule file but for the arm and intake. in robot.py I added the controller inputs needed to control the arm and applied the desired power to the arm motors.
I had a couple questions that I left as comments if you have some free time and don't mind looking through them they are located at:
SwereveModule line 27
SwereveModule line 92
SwereveModule line 107
SwereveModule line 127
SwereveModule line 165
drivetrain line 99
robot.py line 101
and I would appreciate if you could look over all of arm as I wrote all of it